lib verify
