+++ lib/verify.js
@@ -1,0 +1,65 @@
+const sodium = require('sodium-native')
+const request = require('hypercore-signing-request')
+const z32 = require('z32')
+const c = require('compact-encoding')
+
+const { version } = require('../package.json')
+const { Response } = require('./messages')
+const { hash } = require('./secure')
+const { MAX_SUPPORTED_VERSION } = require('./constants')
+
+async function verify (response, signingRequest, pubkey) {
+  if (!response || !signingRequest || !pubkey) {
+    console.log(`hypercore-verify ${version}\n`)
+    console.log('Verify a signed message.')
+    console.log('\nUsage:')
+    console.log('hypercore-verify <response> <signingRequest> <pubkey>\n')
+    throw new Error('Invalid arguments')
+  }
+
+  const res = c.decode(Response, z32.decode(response))
+  const publicKey = z32.decode(pubkey)
+
+  let req = null
+  try {
+    req = request.decode(z32.decode(signingRequest))
+  } catch (e) {
+    console.error('\nCould not decode the signing request. Invalid signing request?')
+    throw new Error('Invalid signing request')
+  }
+
+  if (req.version > MAX_SUPPORTED_VERSION) {
+    throw new Error('Request version not supported, please update')
+  }
+
+  if (Buffer.compare(res.requestHash, hash(z32.decode(signingRequest))) !== 0) {
+    throw new Error('Signature was not made over this request')
+  }
+
+  if (Buffer.compare(res.publicKey, publicKey) !== 0) {
+    throw new Error('Public key does not match')
+  }
+
+  const signables = request.signable(publicKey, req)
+
+  if (signables.length !== res.signatures.length) {
+    throw new Error('Invalid response: signature count does not match')
+  }
+
+  for (let i = 0; i < signables.length; i++) {
+    if (!sodium.crypto_sign_verify_detached(res.signatures[i], signables[i].signable, publicKey)) {
+      throw new Error('Invalid signature!')
+    }
+  }
+
+  console.log('\nSignature verified.')
+  console.log(`\n${z32.encode(publicKey)} has signed the following request:`)
+  console.log({
+    core: req.id,
+    fork: req.fork,
+    length: req.length,
+    treeHash: req.treeHash.toString('hex')
+  })
+}
+
+module.exports = verify
